refactor: global util 패키지 세분화
